Constants: set DEFAULT_MAX_BUFFER_ENTRIES to 1000
We were previously limiting write buffer size only by total number of
bytes & not by number of rows; for imports where many rows would fit
into a given number of bytes, the allocation load for the rows
exhausted the heap. In testing on the laptop, a default of 1000 didn't
tax memory & also did not impede total throughput (the writer threadpool
simply grew). We'll likely need to do testing on the cluster to see if
1000 remains a good default.