Add FindHashJoinColumns, for identification of cols to hash on join
This version (compared to our previous impl on the Derby side) records
columns to hash in terms of the immediate left & right children of the
join, as opposed to table-number & column pairs. The latter required
runtime translation into offsets on immediate children, & failed in the
case of references to more than one underlying table on a single side of
the join.