Add IJoinRowsIterator & an implementation for MSJ
IJoinRowsIterator is half of the approach to separating concerns and
consolidating logic in join execution. IJoinRowsIterators are concerned
with producing left & right rows which may be joined; a future
componenet will consolidate the logic by which they're joined that is
now present in MSJoiner (so that it can be used in NLJ, Broadcast, & others).