FromBaseTable: reduce estimated cost for covering indexes
For a query over a table without any filters, previously the cost
estimate for a table scan was equal to the estimate for a covering index
scan. Though both will read the same number of rows, the rows in the
covering index are smaller (strictly speaking, smaller or equal)
so overall it represents a lower IO cost for Splice. This change
slightly reduces the cost for the covering index to break the tie.