SubqueryNode: pull materializeRS calls up into execute()
Previously, in the generated Activation, calls to
BaseActivation.materializeResultSet() happened just before the subquery
was opened. Given Splice's distributed execution model, this meant that,
depending on the shape of the tree, subqueries could be materialized
multiples times (at each region of a table).

This change pulls the materializeResultSet() calls into the execute()
method so that materialization happens only once, before execution of
the main query.