Add CachedRS to abstract over Derby's union of RowRS
Derby represents an in-memory materialization of a result set as an
(n-1)-way union of RowResultSets. This commit replaces construction of
that union with a call to getResultSetFactory().getCachedResultSet() so
that the exact representation can be overridden in other result set
factories. (Derby's impl of CachedResultSet is the same union of
RowRSs).