OperationRS: only add StatementInfo once per open RS
Previously there was a problem with subqueries that the Derby activation
was not closing after each execution. A new StatementInfo was created
for each execution & added to the StatementManager, then never removed
(since no close()), & the heap was quickly exhausted. We've changed the
Derby activation to call close() on subqueries, but I'm leaving in this
code as a safeguard.