DB-1442 Add IT, fix bug for BCAST over MERGE w/o intervening sink
Merge was changing the shared SpliceRuntimeContext which Broadcast used
as well. The fix here is to make a copy of the SpliceRuntimeContext
object in the Merge join so ancestors won't see changes. But using the
context object to pass the "scan start override" isn't great; I'd rather
use an explicit parameter in executeScan. I'll investigate doing so.