DB-1471, DB-1472 Change joiner initialization idiom not to call open()
Root problem with above bugs was that Merge join was calling open() on
its left resultset, which ended up being the second time open() was
called on it. When a MergeSort join's open() method is called for the
second time, it closes its scanner. So Merge over MergeSort produced
truncated rows for the MergeSort (or an NPE if under a sinking op).

So now new idiom for joiners is to open any resultset that needs
opening before instantiating the joiner; consumers of the joiner itself
then will not call its open() method.