JoinSelector: remove rightEquiJoinColIsPK property
This property was constantly false, so removing it as dead code. The
reason it's gone is because Derby handles the case the property was
intended to help identify before the JoinSelector runs. The property was
around to identify a constant filter on a join column that was the
primary key on the join's RHS table, e.g.

  select * from a, b where a.id = b.id and a.id = 1

However, Derby rewrites this as a cross join equivalent to

  select * from a, b where a.id = 1 and b.id = 1

and for cross joins, the join selector chooses NLJs, the strategy we
wanted for these constant lookups anyway.