DB-1695 - aggregate functions as window function order by column expressions. On the splice side, I just added tests.
Got it to work when the WF is over a GROUP BY, but still failing when GROUP BY
is removed. Seems that we're building a GroupByNode because we're finding the
agg fn inside the WindowFunctionNode.  Need to find a way to filter agg fns
when they're inside WFs, but then we'll likely have to take on the work
GroupByNode was doing to push the agg fn calculation down in the tree so that
the WF only has a CR to the agg result (since we need the result for a key
column).