DB-2124 - NPE when WF is part of a BinaryArithmeticOperatorNode. The NPE was caused by not setting the WrappedAggregateFunctionNode's operator to the the value of its delegate wrapped AggregateNode (was accessed directly rather than thru overridden getter). Fix was to keep it in sync after binding.
The bigger problem, revealed after above fix, was because the WrappedAggregateFN was
under a BinaryArithmeticOperatorNode as one of its operands. In that case, we need
to handle the WF aggregation in the WindowFunctionResultSetNode, replace WF with a
reference to the aggregate result and push the BAON into the projection where the
BAON can perform its operation with the numeric values computed by the aggregation.