Revert Fix to DB-2068 and DB-2191
Revert "Fixing compilation errors caused by reverting changes"

This reverts commit 2bf53b2db36dfd88f5fdbe936a252b69351b770a.

Revert "Revert "DB-2068: Making it so that Subqueries flatten when there are no predicates in the overall subquery restriction""

This reverts commit 6adab6e668c18022bbb246cdf59617c70810e9dc.

Revert "Revert "DB-2068: Made it so that at least NestedLoopJoin always works with a Union on the right-hand side (e.g. with a flattened subquery)""

This reverts commit 180d517c8f00ab0430209212bd4c478fbb25e56b.

Revert "DB-2191: Adding a method so that we can fix MergeJoin behavior"

This reverts commit a3e14d6f16f4373134223b01d13665edda879e23.

Revert "DB-2191: Making it possible to unmark a predicate. This allows us to make a predicate no longer be considered a start/stop key or qualifier"

This reverts commit c1903fa83a0ef6dd023a417a2aca7b639036e7bd.

Revert "DB-2068: Made it so that at least NestedLoopJoin always works with a Union on the right-hand side (e.g. with a flattened subquery)"

This reverts commit a3291420d6fa10ee83b9c489b1dfd867e99b615a.

Revert "DB-2068: Making it so that Subqueries flatten when there are no predicates in the overall subquery restriction"

This reverts commit 8a33fcf8aa215ca81c43ab7f5f4c7260b356814b.

Conflicts:
	java/engine/org/apache/derby/impl/sql/compile/PredicateList.java
	java/engine/org/apache/derby/impl/sql/compile/ResultColumnList.java