DB-2033 - data type overflow ("outside the range") for window function aggregate.
When we aggregate a column (addend), we clone it to store the first result. From
then on, the result column is the type of the addend, not necessarily the result
type.  This results in an error that's trapped when a new result would overflow
the type of the column.

The fix was to pass in the result type when creating the window function and
use the cloned addend only as the seed and use its value when setting the function
result.