DB-2261 - window function query returns 0 rows.
This commit addresses an NPE found only in a clustered setup while
testing these queries.  The NPE happened reliably in
MergeJoinOperation.updateStats(), where the joiner was null in this
case.
The fix was simply to wrap an if not null guard around joiner and
rightRows in that method before using them to build stats.