DB-2290 - Sum window function on column with NULL values gets wrong answer.
Because Derby's SQLDecimal "nulls out" the result for plus() (and minus())
when an input value is NULL, we were essentially starting the summation
over when we hit a NULL input value.

The fix was to skip the summation, and setting the result when we encounter
a NULL input value.  This is essentially the same as summing the previous
result with zero but without doing the work.  We now get the same result as
Oracle and PostgreSQL (tested in sqlfiddle).

Note that I made the same change to SumAggregator#calculateOnAdd() and
SumAggregator#calculateOnRemove() -- it uses SQLDecimal#minus().
I also made the same changes to AvgAggregator, as it uses plus() and minus().
Added test verification.