DB-2398: Query timeout. Derby Statement was not setting query timeout on the Connection.
The fix was in Derby to allow the Derby Statement to set the query
timeout set by the client on the Derby client Connection, which sets
the timeout on the Socket.

Added tests to assure query timeout rolled back an update that timed
out. Tests required creating a stored procedure that would sleep a
configured amount of time imitating a lag in the update.

Had to modify SpliceWatcher to check if the Connection was closed
when deciding to get or create the Connection.