DB-2472: rank() function buffer underflow.
The problem was in WindowFunctionNode where we were pulling the ORDER BY column
up from below a join node.  The plan tree looked correct but the node reference
was pointing "beyond the grave".  This caused the column reference to point to
a column of a different type during the splice scan.

Added source column result set in the splice PlanPrinter to help diagnose this.
When the plan is printed with TRACE, the result column information is printed
and now column reference source result set number is printed.  Use
call SYSCS_UTIL.SYSCS_SET_LOGGER_LEVEL('com.splicemachine.derby.impl.ast.PlanPrinter', 'TRACE');
to log this information.