DB-2536: NPE in rank() Merkle query 1.
The query is a behemoth with several, several joins. It seems like the join
and coalesce nodes were confusing the derby window function planner, which
has to search the plan below for any nodes missing from the select clause,
create references and pull those up into the window result set node.

The fix was to add extra null guards around places where the code tried to
get sub expressions or result set nodes from a given node (because derby
plan nodes are inconsistent in their method naming and implementation [just
Derby bitching here]).

Although I got the original query to run without throwing an NPE, I was unable
to reduce neither the data set nor the query down to something that's still
reproducible but able to be run as a Unit Test. That would have been optimal
because it would have stayed with the code and been regression tested.  I did,
however, attach the script bundle, as well as the successful query plan tree,
to DB-2536.