DB-2863: second(timestamp) function fails.
The problem originated when we added the Lazy*DataValueDescriptor classes.
SQLTimestamp.getSeconds(NumberDataValue source) enforces that the 'source'
is of type SQLDouble. The 'source' type is now LazyTimestampDataValueDescriptor.
The fix was to change the enforcement so that the NumberDataValue checks that
it contains a type double and all went fine. I ran the ITs on 1.0.x and all passed.

Added a simple IT, LazyTimestampDataValueDescriptorTest.