DB-1596: truncate decimal
Found the problem with returned precision and scale. Code in
TruncateOperatorNode was using default (built-in) type for leftOperand
(the decimal to be truncated) which sets the decimal precision to 31
and scale to 31 (the max). Setting precision and scale to the same
value causes a calculation in DDMWriter to think its an overflow.

The fix was to set the node's type using leftOperand's DataTypeDescriptor
which is known after bind time for both literal values and column
references.