DB-2871: LazyTimestampDataValueDescriptor ClassCastException.
The problem was that Derby's SQLTimestamp.timestampAdd() method was casting LazyTimestampDataValueDescriptor
to a SQLTimestamp. LazyTimestampDataValueDescriptor was added to wrap a SQLTimestamp for efficiency.

The fix was to pull the wrapped SQLTimestamp out of LazyTimestampDataValueDescriptor before
it forwards the execution to SQLTimestamp for timestampAdd().  I checked for other places
a cast to SQLTimestamp is made in Derby's SQLTimestamp and added tests for those.

Corrected faulty SpliceDateFunctionsIT and added tests.

Minor modifications made to SimpleDateArithmeticIT and TruncateFunctionIT.