DB-1893: drop column not transactionally correct.
Found and fixed an NPE in DataDictionaryUtils.getTableVersion(). Happened when table
was dropped or created in another txn not visible to current. It now returns null
when it can't find a TableDescriptor.  Only used by RowTransformer.initEncoder() to
get the table's DescriptorSerializers for the given version.  When null, it returns
the latest version, which seems appropriate when table is just dropped or created.

Unignored test in DropColumnTransactionIT.