DB-2705: alter table add check constraint throws NPE
The problem was that there was a bunch of state in AlterTableConstantOperation that wasn't being
initialized since ModifyColumnConstantOperation was created as a sub class. Some of this state was null.

Removed the state and added matching arguments to the methods that used it.
Removed some unused methods.
Check constraints still don't get checked but no NPE.
Added an IT.