Adding a Transactions test category.
The purpose of this is twofold: First, it makes it easier for us to find
all the tests which are transactional in nature: just do <egrep
'Transactions.class' -R *> from the splice_machine_test directory.
Secondly, it will allow us to (eventually) categorize which tests run
and in what order, and (hopefully) to speed up test execution.