DB-1755: alter table. Alter table actions change conglomerate descriptors in TableDescriptors. Doing this on main table conglomerate descriptors causes TableDescriptor's cached "heapConglomerateNumber" to become invalid.
Invalidating this cached conglom num would be simple if conglomerate
descriptors were modified through this object itself (as though it were
the single source of truth for table metadata, which is how it's used in
many places), but callers get the conglomerate list and modify it directly.

Given that, I opted to invalidate the cached conglom num whenever the
conglom list or one of its members is requested. In fact, I could have
modified all users of this object to add/remove congom descriptors using
this object's interface but that would have made much larger impact and
would only handle part of this object's problems with cached table
metadata.