DB-1755: alter table, add PK. Found that populate task transformer was incorrectly decoding source rows with the target row's PK column row key.
Encoders/Decoders must, of course, both be created with a key encoder/decoder
specific to whether a key exists on one side or the other. Adding a PK is an
asymmetric operation w/r/t the source and target tables - the source has no PK
but the target does (ignoring the fact that other keys could be present).

Given that, we need to track column ordering for both sides (src and target) and
supply that info to the alter table descriptor that creates the decoder/encoder.

It seems obvious when you verbalize it but it's not when you look at decoder/encoder
creation.  Much of this dial twisting and lever pulling should be made internal to
data encoding.  We'd have less errors I think.