DB-3181: Fixing NPE in BaseDistribution.
It turns out that an entire partition can consist of null elements. When
that happens, we have an error that's thrown by Object distributions. To
avoid that situation, we check for a null partition range. When that
happens, we return the null count only if the entire range of request is
null(as that matches the SQL spec behavior).