Changing explain cost info for better readability.
This changes the confusing "localCost" and "remoteCost" to
"processingCost" and "networkCost" when displayed in ExplainTrace. This
makes the distinction clear about where costs are determined--the cost
to process might be considerably more than the cost to read the final
data over the network (or vice versa). As a result, things should be a
lot easier to interpret.