DB-3228: Removing StackOverflowError.
The query optimizer in DML049 considers 9!*4 different plans,
which is a lot. We weren't clearint out the base cost estimates when
it reset from each permutation check, so you ended up needing to parse
through 9!*4 Base CostEstimates in a recursive stack, so you ended up
blowing the stack.

This clears the Base cost in between each new join order, so that
we end up without the bad costing