DB-2887: Ensuring that index scan and lookup lists are separate.
When performing a non-covering index scan, we need to ensure that the
cost to read the base data is accurate. This requires us to correctly
know which columns are being fetched from the base table, and which are
present in the index. This ensures that the two lists are separate and
well configured.