Representing cost in milliseconds during explain.
The base "units" for the costing mechanism is microseconds (really, it's
not important what the units are, only the relative differences, but
sometimes it's nice to know the units). The problem here is that large
numbers are hard to read; by shifting the display unit down to
milliseconds, we get a much more readable cost number that doesn't lose
any fidelity since it's only for display. Of course, costs < 1
microsecond will not be well displayed, and heap sizes < 0.001 MB will
display as 0, but since 0.001MB is just a single KB, it doesn't really
make any difference.