DB-3194: Allowing maximum width to be disabled.
Explain plans which contain many nodes and/or lots of predicates on a single node don't display well with a fixed width. There are other situations which display similar characteristics (1 column string-only tables, for example). In those circumstances, it's convenient for us to have the ability to disable maximum widths. This commit makes it possible to disable maximumdisplay with by setting MAXIMUMDISPLAYWIDTH 0; in ij.

This affects no other logic, only the display of strings, and it's not enabled by default--a user must explicitely call MAXIMUMDISPLAYWIDTH 0; in ij for this to take effect.