DB-2887: Better index statistics management.
Index statistics are somewhat tricky. We need to include both the cost
to perform a multi-get, and the cost to perform an actual scan, and we
need to do so without measuring both values, since we don't want to
change the storage model for statistics. To do this, we use a "ratio
adjustment", where we use the remote scan latency for the base table,
but we multiply it by an adjustment factor which is the ratio of the
average row width of the index to the average row width of the base
table. Then, when estimating the lookup cost, we use the "multiGet"
latency, which is just that which is stored in the statistics table for
the index itself. We wrap all this up in a few nice object abstractions
to clarify how access happens and to match the API specifications.