DB-2887. Apply index lookup costs before non-qualifier predicates.
Derby checks non-qualifier predicates in a ProjectRestrict which sits
above the relevant tablescan/index lookup. Thus, when we cost an index
lookup, it is important that the index lookup cost as measured should
reflect the fact that the lookups occur BEFORE the non-qualifier
predicates are applied. Otherwise, we underestimate the number of rows
which are fetched during the index lookup stage, and therefore will
favor indices when they shouldn't be.