DB-3287. Better MergeJoin feasibility rules.
The MergeJoin feasibility rules were very needlessly restrictive, and
also poorly tested. This commit adds a bunch of tests around feasibility
and correctness for two-table merge joins, and the relevent fixes to
MergeJoinStrategy to ensure all tests pass. Room for more rules and
tests as we find new scenarios where MergeJoin is allowed to be correct.