DB-2001. Removing infinite loops for infeasible join strategies.
When a join strategy is infeasible, but hinted, Derby wants to throw an
error. Unfortunately, it throws the error only *after* the entire plan
has been optimized. Since a subtree can't be optimized because it has an
infeasible plan, the tree cannot be optimized so it falls into an
infinite loop of optimizing attempts. The fix then is to verify that a
best plan has been found after every subtree has been optimized (in
particular, SelectNode and TableOperatorNode, since those are the two
mentioned in DB-2001 and DB-2877).