DB-3155. Making IN clauses cost appropriately
Derby, as it turns out, has a large number of complicated and confusing
approaches to costing various scan predicates. As a result, the
refactoring to include statistics costing in the optimizer failed to
take IN clauses into account. Investigating, we see that Derby didn't
really take them into account either, essentially just deciding to pick
arbitrary values for the costs. Thus, in order to make it work, we have
to significantly refactor the Store costing strategies to take into
account probe clauses in the first entry, which is now internal to
ScanCostFunction.java, where all scan costing logic is contained.