DB-3238: Making hinted Broadcast joins ignore heap size limits.
When unhinted, we use heap size limits to determine whether or not
Broadcast join is a feasible join strategy. When the query is hinted,
however, we now assume that the user knows what he's doing and allow it
through as long as it is logically correct (i.e. a hash join), even if
the optimizer believes that the join would result in an OutOfMemory
error. A side-effect of this is that it will be possible for users to
CAUSE OutOfMemory errors if they force a broadcast join when the right
side is too large.