Using RegionLoad information when statistics are not available.
When no stats are available, we use the RetgionLoad information provided
by HBase to give us an estimate as to the average row width and the
number of rows. This is pretty terrible (it lacks lots of important
statistics), but it's better than nothing.

Making Region Loads properly estimate based on sizes, and making BroadcastJoin properly understand singleScanRowCount