DB-3354: Nested triggers produce null transition variables.
We have a table T1 that has an AFTER INSERT ROW trigger TEC1 defined on it (fires when a row gets inserted into T1).
TEC1 inserts a row into a table, T2, which has an AFTER INSERT ROW trigger TEC2 defined on it (fires when a row gets
inserted into T2), which inserts row into table T3.

The problem we're seeing with these changing triggers is that we're getting the wrong trigger execution context when
we fire a trigger. Specifically, we create TEC1 (which has a non-null transition row because the insert kicked off
the chain) and push it onto the trigger execution stack, then we create TEC2 (which has a null transition row because
TEC1 has not fired yet and so no row was inserted into T2) and push it onto the stack. TEC2 is now at the top of the
stack. Then we fire TEC1 and pop a TEC off the stack to execute but it's the wrong one; it's TEC2 which has a null
transition and so null gets inserted.

Moved TEC push to LCC closer to when the trigger gets fired in TriggerEventActivator.  Push only once; pop in cleanup().