DB-3421: Fixing sort order error.
For extremely negative longs (numbers which are negative, and which
require a full 8 bytes to represent) the header byte was getting messed
up. By completely random luck, this didn't cause the encoding to break
(i.e. encoding the number, then decoding it again would result in the
same number), but it *did* result in and improper sort order with
respect to less negative numbers.

This patch fixes the underlying sort order issue, and also adds a
significant systemic test which attempts to cover a large enough
subspace of numbers to ensure that this does not occur with other
numbers.

Also adding additional tests for BigDecimal encodings to facilitate
refactoring, and adding some refactoring of ScalarEncoding to make sure
that the tricky bits are as encapsulated as possible.