Added optimizer trace in and around JoinNode targeted to predicate movement.
Interestingly, in spite of the fact there's an optimizer field in parent ResultSetNode,
which is accessible from subclasses, the reference is null and the only reliable ref you
can get a tracer from is the Optimizer passed in to optimize() methods.

Since Derby plan optimization happens, and predicates are moved, in many other places
besides optimize(), the tracer needs to be refactored to not have an Optimizer dependency.