DB-3490: Plan printer/debugger NPE when printing plans after parse phase.
PlanPrinter and PlanDebugger (light-weight printer) were both attempting to access the
plan for costing information to print. Costing information has not been calculated at
this point and was null resulting in NPEs.

I added guards so that costing information would not be accessed in the after parse
phase and, if it was null, to not attempt to dereference it.