DB-3517: Correcting AIOOB error.
It turns out that start and stop key array lengths can differ while not
being null. This is weird, but manageable. We treat running off the end
of an array as a non-existent requirement, which is effectively >/>=
when running off the stop key, and </<= when running off the start key.