Making build more robust.
The TrafficShaping code isn't used anywhere, but in case it ever does,we
want these tests to be robust. I added a specific Clock handler in place
to make sure that we can do this. The code itself is fine, we just
needed tests which were not as prone to timing issues.