DB-3474: Deduplicating buffer on first merge.
When merging the buffer into a sparse array for the first time, it is
necessary to remove any duplicates that the buffer might contain, and to
strip zeros from the computation. Otherwise, you end up with an
incorrect buffer estimation for very low cardinalities.

This checkin resolves that issue, and adds unit tests for specific cases
(not a very wide spread, but it does cover the scenario of interest).

Adding documentation to SparseHyperLogLog