DB-3474: Correcting configuration.
There were two situations: First, the accuracy of the cardinality
estimator was low by default (8); this fix adjusts that to 14. It will
use more memory for high cardinality data sets, but will be considerably
more accurate in general (0.008 instead of 0.065), and will use the
sparse form for smaller cardinality data sets more frequently (up to
around 24K).

Secondly, there was a mental typo, where the precision factor for
frequent elements and the precision factor for cardinality were reversed
upon construction, resulting in a cardinality precision of 5 in
practice, instead of the configured value.

This commit resolves both issues. Since one is a change of
configuration, and one is a literal one-line change, no unit tests were
added. Instead, unit tests in Statistics were added to ensure that, when
properly configured, the algorithm will be correct for the DB-3474
scenario.