DB-3750: trigger event referencing old transition values instead of new (reopened)
The problem here was that two identical triggers were being fired for the same event
but the event was being cleared from the TriggerExecutionContext (TEC) in between and
not being reset.

A TEC is a container of trigger information (definition, event, triggering resultSet,
etc) which gets used each time a trigger fires. These containers get reused for possibly
different events and definitions and certainly different triggering results.

The fix was to move the setting of the TEC's event to where its definition was being set
and cleaning up of the TEC cleanup logic.