Adding a ConcurrentTicker.
There are locations in other projects (specifically in Misery Maker)
which could make use of the Clock abstraction introduced to make testing
concurrent time-based utilities more convenient. To this end, we
introduce the ConcurrentTicker, which allows us to have deterministic
clock behavior while remaining thread-safe(at least, for a given value
of "deterministic" in multi-threaded applications).