Fixed DB-3672.
The Batch size in Asynchbase was 4096, so updates to the same row that
exceed that number of KeyValues (2048 updates) will result in a single
batch pulling less than the entire row. We correct this in the short
term by making the Asynchbase batch size match the value set in HBase
scan creation (-1), which means read all the versions for all columns.
This is tricky from a memory perspective, since it can lead to OOM
errors with a lot of updates, but carrying through the HBase batch size
to asynchbase is good anyway; when we set the batch size at a later
date that batch size will be carried through.