DB-3656: add support for generated columns back in.
Derby and Splice were doing everything correctly to support this except
the GeneratedMethod was being lost in DMLWriteOperation, super class of
InsertOperation and UpdateOperation.

All that was needed was to track the GeneratedMethod in SpliceMethod by
the method name, then invoke the method on the activation for each row.

Added ITs.