DB-3688. Corrects NPE in merge code.
The merge code has 4 distinct states
(Dense,Dense),(Sparse,Sparse),(Sparse,Dense), and (Dense,Sparse). These
all need to return the correct results, and that required careful
investigation and Unit testing, which Scott was lazy and didn't want to
do. This commit adds those tests, and corrects the inevitable problems
which occurred as a result of the lack of coverage.