DB-3705: Dealing with no-uniform histograms.
When the entire histogram can fit within the frequent elements, then we
have a situation where using the base linear estimation will result in
massive numbers (which don't make sense). The commit adjusts those
values and ensures that in that circumstance exact answers are returned.
To confirm, UTs are added.