DB-3729: Correcting bounds errors in distributions.
An invariant of Distributions is that the range selectivity of a given
interval of data should never exceed the total number of rows recorded
in the system. This invariant was being violated in
UniformtStringDistribution in two major ways: the utf8Position
computation did not properly compute the utf8 position (it did not
increment the character count properly), and the linear expression used
did not use the proper maximum row count. Because the linear expression
is only used for the "uniform" part of the histogram (i.e. NOT the
frequent elements portions), we cannot use the total nonNullCount(), we
must use the adjusted row count instead.

The first problem is unique to strings, but the second problem occurs in
all data types, so this commit also corrects the issue in all data
types, and adds tests for those conditions. In the course of testing, we
discovered several issues caused by data type overflows during range
selectivity estimation, which we also corrected. Those in turn led us to
consider the complications due to floating point NaN, +- Infinity, and
other boundary conditions, which we also tested and ensured proper
behavior for in this commit.