DB-3735. Correcting all-null updates to indices.
Indices write rows containing nulls slightly differently. There was a
slight mental typo in IndexTransformer that resulted in entries being
skipped if they were null. This is the "sparse" representation. However,
we cannot use the sparse representation of nulls for row keys (which is
all that matters with indices), because otherwise we won't know when one
field ends and another begins. Thus, DB-3735 was caused by the reader
intrepreting bytes of the row key as the second entry in the index.

This fixes the mental typo, and adds data-type coverage ITs for all 1
and 2-column indices to ensure that all data types properly encode and
decode. Because this fix affects on-disk storage, indices built prior to
this fix which contain null entries will need to be rebuilt to ensure
the correct encoding and decoding.