DB-3806: Refactoring DDL mechanism and adding UTs.
We want to be able to unit test the DDL coordination mechanism, to
ensure that our semantics are clean and coherent and (if possible) to
detect any race conditions based on order of operations. This can be
done if we carefully restructure the code.

Here we refactor the code into several subcomponents, which we can write
simulated tests for independently, as well as together. These simulated
tests do not use ZooKeeper, but instead use internal test-datastructures
to simulate zookeeper behavior, in such a way that we can play with the
sequence of events for each component.