DB-3813:Trim trailing whitespace for varchar stats.
In Splicemachine, "select * where varchar = 'K'" and "select * where
varchar = 'K '" return the same results. However, statistics were making
a distinction between the two possible values. This commit eliminates
the ambiguity by trimming whitespace from strings during collection, and
also trimming whitespace during estimation. It adds lines to a
ScanSelectivityIT to verify the fix.

During testing for this fix, we found DB-3836; to avoid interfering with
this bug fix, the lines which fail the test were noted and commented out
temporarily, so that it can be fixed when DB-3836 is resolved, without
the need for new ITs for that as well.