DB-4040: Correcting startup race condition.
It's possible for multiple simultaneously starting RegionServers to both
attempt the upgrade simultaneously. Our concurrency controls will force
one to win and the other to lose; in that case, we were attempting to
add a column which had just been created by another server. This isn't
actually a problem, but it will throw an exception. Since all we care
about is that it was created (and not who created it) the exception is
as good as success here.