DB-4043: Use reference counting to clear Broadcast Join cache.
The BroadcastJoin cache was using a Google guava Cache to save on scans
against the inner table, which is fine as far as it goes. Unfortunately,
the guava cache does not immediately invalidate values, instead keeping
them around for some time. As a result, the garbage collector was unable
to collect right hand sides efficiently, which ended up resulting in an
OOM error during execution.

This adds in reference counting to the inner hash table structure, which
causes those references to immediately be released, helping the garbage
collector keep up.

Unfortunately, this version still has a very high amount of garbage
churn, due to the byte arrays and byte buffer generation during key
lookup, and because of ExecRow cloning of values. More work will need to
be done for that to resolve itself.