DB-4043: Making lazy DVDs actually be lazy.
The idea behind Lazy DataValueDescriptors is that you don't decode data
unless you actually need the decoded data. This is not necessary when
performing comparisons (since our encodings preserve ordering), or when
computing equality checks (since our encodings are unambiguous).
However, the existing Lazy DVDs were effectively deserializing entities
when they weren't necessary. In addition, they were caching an internal
reference to a DataValueDescriptor for the decode values. This internal
reference is useful in that it avoids us having to recreate that logic
in another location, but is problematic because it would be cloned along
with the lazy elements.

This commit refactors the Lazy DVD structure so as to be A) actually
lazy for equality and comparison checks, and B) to avoid excessive
memory usage due to extra internal object references.

There were also various small errors in the LazyDataValueDescriptor logic
that we had inadequate testing around. This is a step in the direction
of fixing that.