DB-4043: Closing JoinTable more carefully.
This ensures that the JoinTable is removed on closure of the
BroadcastJoin(so that if there is an error, we still clean up our
references). There was a minor observable performance optimization that
occurs when we close it after the last row is found out of the joiner,
so we left that in with a note as well (this results in ~16% performance
improvement for one test, for example).