DB-358 DB-4115
Added code for LazyDate and tests for LazyDate and LazyTimestamp
in spliceengine modeled after LazVarchar.

- In order to avoid having to bring in all of Derby with a DBContext,
added a flag to skip DB Context.
- Fixed bug DB-4115 DateTime did not deserialize correctly
when it was called when there was no calendar ,as per Scott
suggestion just changed the code to use a new instance of Calendar
and execute the other path of the code.