DB-4344: 'badRecordsAllowed' parameter has off by one error Was in SparkPairDataSet#insertData(). Checking (badRecords.size() >= insertOperation.failBadRecordCount). Changed to (badRecords.size() > insertOperation.failBadRecordCount).
DB-4345: Error status records are appended to 'bad' record file with no separator
It was in SparkPairDataSet. If the the badRecords List<String> is non-empty, appending System.lineSeparator() to each non-empty element.
I added it here so that any perf penalty acquired would be in the context of an error condition, minimizing impact.