Manual integration of f618209 from master_dataset to k2_refactor:
DB-4170: Create table with data.
Removed creation of InsertNode, binding and optimizing of if in CreateTableConstantOperation
to pure execution of query string as an insert in CreateTableConstantOperation.
The problem was that binding and optimizing alone of the InsertNode was not enough. When
executing a query with more than 2 joins, the join predicates were not being found.
This method takes the query through the full parse/bind/optimize/generate steps after the
required table creation.