DB-4669: Avoiding over- and under-insertion problems.
When performing "insert into foo select * from foo", we are reading
from the same table that we are inserting into. It turns out that HBase
has a weakly-consistent iterator policy, so some of those insertions
could end up being seen by the same scanner. If the rows written by that
insertion are then seen by its same scanner within the same transaction,
then the rows are visible and therefore are scanned (and hence
inserted).

To avoid this problem, we wrap the writer itself in a child transaction
of the operation performing the write. This avoids the weakly-consistent
iteration problem and therefore avoids over-inserting records.

There were two issues which results in under-insertions:

1. We were not properly rejecting some bulk writes which are directed to
absent regions (either NotServing, or WrongRegion). We were just
ignoring them.

2. We were not iterating properly through the collection of BulkWrites,
so some rows would just go missing.

This commit addresses both issues, and adds some trace-level logging for
tracking some of these issues.